transaction_record_map to new file
